+++ src/lib.rs
@@ -1,11 +1,13 @@
 pub mod frame;
 
-use num_enum::TryFromPrimitive;
-use serde::Deserialize;
+use num_enum::{IntoPrimitive, TryFromPrimitive};
+use serde::{Deserialize, Serialize};
 
 /// Numeric representation of the message types.
-#[derive(Clone, Copy, Debug, Deserialize, TryFromPrimitive)]
-#[serde(try_from = "u32")]
+#[derive(
+    Clone, Copy, Debug, Deserialize, IntoPrimitive, PartialEq, Serialize, TryFromPrimitive,
+)]
+#[serde(into = "u32", try_from = "u32")]
 #[repr(u32)]
 pub enum RequestCode {
     /// Start DNS filtering
@@ -31,4 +33,20 @@ pub enum RequestCode {
 
     /// Update the list of trusted domains
     TrustedDomainList = 0,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::RequestCode;
+
+    #[test]
+    fn serde() {
+        let serialized = bincode::serialize(&RequestCode::ResourceUsage).unwrap();
+        assert_eq!(
+            u32::from_le_bytes(serialized.clone().try_into().expect("4 bytes")),
+            RequestCode::ResourceUsage.into()
+        );
+        let code = bincode::deserialize::<u32>(&serialized).unwrap();
+        assert_eq!(code, RequestCode::ResourceUsage.into());
+    }
 }