+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* `RequestCode` supports serialization with serde.
+
 ## [0.2.2] - 2022-05-16
 
 ### Added
@@ -42,6 +48,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
+[Unreleased]: https://github.com/petabi/oinq/compare/0.2.2...main
 [0.2.2]: https://github.com/petabi/oinq/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/petabi/oinq/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/petabi/oinq/compare/0.1.1...0.2.0